Range should work with integers.
A range of integers makes much more sense. The conversion to integers
also implies that the iteration over the range is simpler.

+++ src/main/java/net/cilib/predef/Predef.java
@@ -38,7 +38,7 @@ public final class Predef {
         return new Array<A>(list.toArray());
     }
 
-    public static Range range(double start, double end) {
+    public static Range range(int start, int end) {
         return Range.of(start, end);
     }
 }
